Expose extraction score diagnostics on detail screen
